share working component add carousel
